Added back in some of my old functionality around the menu bar at the top of the full screen files.
Added in a dropdown to the action menu that can do download and local share of files.  The download is setup to handle many files being able to be submitted to it in the background, and maintain the user knowing about current downloads by reporting them in a notifications system.